moved config schemas
