+++ mkdocs_multirepo_plugin/structure.py
@@ -3,7 +3,6 @@ import asyncio
 import os
 import shutil
 import time
-from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Callable, Dict, List, Optional, Tuple, Union
 
@@ -21,36 +20,6 @@ from .util import (
     log,
     remove_parents,
 )
-
-
-@dataclass
-class RepoConfig:
-    section: str
-    import_url: str
-    section_path: Optional[str] = None
-
-
-@dataclass
-class NavRepoConfig:
-    name: str
-    import_url: str
-    imports: List[str] = field(default_factory=list)
-
-
-@dataclass
-class MultirepoConfig:
-    cleanup: bool
-    repos: List[RepoConfig]
-    nav_repos: List[NavRepoConfig]
-    imported_repo: bool
-    temp_dir: str
-    keep_docs_dir: bool
-    section_name: str
-    url: Optional[str] = None
-    dirs: Optional[List[str]] = None
-    custom_dir: Optional[str] = None
-    yml_file: Optional[str] = None
-    branch: Optional[str] = None
 
 
 def is_yaml_file(file: File) -> bool:
